+++ cmd/vayupress/main.go
@@ -554,6 +554,7 @@ func main() {
 	dbpkg.StartWALCheckpointGoroutine(queue.DoneCh)
 	dbpkg.StartStuckJobReaper(queue.DoneCh)
 	dbpkg.StartJobRetentionSweeper(queue.DoneCh)
+	dbpkg.StartArticleTagsBackfill(queue.DoneCh)
 	a.startMetricsSnapshotCollector()
 	a.startSearchReconciler(queue.DoneCh)
 	a.startScheduler(queue.DoneCh)